refactor: Clean up imports and improve code formatting across multiple components
