+++ src/screens/public/PartnerProfile/ServicosContent.tsx
@@ -7,6 +7,7 @@ import {
   TouchableOpacity,
   Modal,
   Pressable,
+  Platform,
 } from 'react-native';
 import colors from '@theme/colors';
 
@@ -188,7 +189,20 @@ const styles = StyleSheet.create({
     padding: 16,
     borderRadius: 8,
     marginBottom: 12,
-    elevation: 2,
+    ...Platform.select({
+      ios: {
+        shadowColor: '#000',
+        shadowOffset: { width: 0, height: 1 },
+        shadowOpacity: 0.15,
+        shadowRadius: 2,
+      },
+      android: {
+        elevation: 2,
+      },
+      web: {
+        boxShadow: `0px 1px 2px rgba(0, 0, 0, 0.15)`,
+      },
+    }),
   },
   serviceName: {
     fontSize: 16,
